refactor: move meta title generation to function file
